[chore/#230] GameMapper 분리
